requirements.txt
altgraph==0.17.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.1
fonttools==4.51.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.7.5
mysql-connector-python==8.4.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pefile==2023.2.7
pillow==10.3.0
pycparser==2.22
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
XlsxWriter==3.2.0
zipp==3.18.2
